requirements.txt
pymongo==4.0.2
scrapy==2.6.1
bs4==0.0.1
git+https://github.com/davvid/black.git@single-quotes
python-dotenv==0.20.0